Refactor sys utils into modules and add CI build
