+++ src/commands/emmc.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+void handleEmmc(const std::string &command, bool debugEnabled);
