+++ src/commands/resize.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+bool runResize(const std::string &partitionUuid, const std::string &partitionNumber,
+               bool forceResizeRequest = false);
