+++ src/ui/banner.cpp
@@ -1,0 +1,15 @@
+#include "banner.h"
+
+#include <iostream>
+
+void printBanner() {
+    std::cout << " #######  ########  ######## ##    ## ##     ## ########\n"
+              << "##     ## ##     ## ##       ###   ## ##     ## ##     ##\n"
+              << "##     ## ##     ## ##       ####  ## ##     ## ##     ##\n"
+              << "##     ## ########  ######   ## ## ## ######### ##     ##\n"
+              << "##     ## ##        ##       ##  #### ##     ## ##     ##\n"
+              << "##     ## ##        ##       ##   ### ##     ## ##     ##\n"
+              << " #######  ##        ######## ##    ## ##     ## ########\n"
+              << "----------------------- SysUtils  -----------------------\n\n"
+              << "Started!\n\n";
+}
