+++ src/ui/banner.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void printBanner();
