+++ src/utils/string_utils.cpp
@@ -1,0 +1,11 @@
+#include "string_utils.h"
+
+#include <algorithm>
+#include <cctype>
+
+std::string trim(std::string value) {
+    auto not_space = [](unsigned char c) { return !std::isspace(c); };
+    value.erase(value.begin(), std::find_if(value.begin(), value.end(), not_space));
+    value.erase(std::find_if(value.rbegin(), value.rend(), not_space).base(), value.end());
+    return value;
+}
