+++ src/utils/string_utils.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+std::string trim(std::string value);
